Refactor shell code in testsuite
- Create shell instance only for the command execute
- Remove channel dependency to get the exitCode
- Use go capability to get exitcode for a process instead running a custom command
- Refactor shell struct configuration

Current implementation start the shell as as backend process and then every command
which is pass to Execute function run it on that backend process. This backend shell
process connected to stdPipe/inPipe and errPipe, so the command execution process is
something following.

ShellInstance() => Execute(<cmd>) => shellProcess.inPipe(<cmd>) => Use goroutine to read std/stderr
=> Channel to communicate the exitCode of the command from that goroutine. => ShellCleanup()

What was happening in above process that end to end there is only one shell instance
and if some command hang during the execution from the inPipe it will block rest of
execution.

With this patch following is expected.
Execute(<cmd>) => ShellInstance => Run the command => ShellCleanup()

We now create shell instance, put the command as stdin and
cmd.Run() take care of cleanup of that shell instance.
This way if command hang then it doesn't block the next execution.