Revert "Refactor shell code in testsuite"
This reverts commit b1002c1e583dc0ca23f1bbb79b1dbcc5eec5e827.

During the use of this commit on crc integration test, we found out
that we do need to depend on single shell instance so easily can able
to replicate the user scenario.

A user usually continue in single shell session, it can be exported some
variable or creating other variables and using those in next commands.
With the current refactor every shell instance is a seperate one for execution
of a command.